+++ es40/src/Keyboard.cpp
@@ -29,6 +29,9 @@
  *
  * $Id$
  *
+ * X-1.3        Brian Wheeler                                   27-FEB-2008
+ *      Avoid compiler warnings.
+ *
  * X-1.2        David Leonard                                   20-FEB-2008
  *      Avoid 'Xlib: unexpected async reply' errors on Linux/Unix/BSD's by
  *      adding some thread interlocking.
@@ -1704,7 +1707,7 @@ int CKeyboard::SaveState(FILE *f)
   fwrite(&ss,sizeof(long),1,f);
   fwrite(&state,sizeof(state),1,f);
   fwrite(&kb_magic2,sizeof(u32),1,f);
-  printf("kbc: %d bytes saved.\n",ss);
+  printf("kbc: %d bytes saved.\n",(int)ss);
   return 0;
 }
 
@@ -1762,7 +1765,7 @@ int CKeyboard::RestoreState(FILE *f)
     return -1;
   }
 
-  printf("kbc: %d bytes restored.\n",ss);
+  printf("kbc: %d bytes restored.\n",(int)ss);
   return 0;
 }
