+++ es40/src/S3Trio64.cpp
@@ -29,6 +29,9 @@
  *
  * $Id$
  *
+ * X-1.12       Brian Wheeler                                   27-FEB-2008
+ *      Avoid compiler warnings.
+ *
  * X-1.11       Fang Zhe                                        08-JAN-2008
  *      Endianess.
  *
@@ -489,7 +492,7 @@ int CS3Trio64::SaveState(FILE *f)
   fwrite(&ss,sizeof(long),1,f);
   fwrite(&state,sizeof(state),1,f);
   fwrite(&s3_magic2,sizeof(u32),1,f);
-  printf("%s: %d bytes saved.\n",devid_string,ss);
+  printf("%s: %d bytes saved.\n",devid_string,(int)ss);
   return 0;
 }
 
@@ -551,7 +554,7 @@ int CS3Trio64::RestoreState(FILE *f)
     return -1;
   }
 
-  printf("%s: %d bytes restored.\n",devid_string,ss);
+  printf("%s: %d bytes restored.\n",devid_string,(int)ss);
   return 0;
 }
